requirements.txt
requests
fastapi
uvicorn
psycopg2-binary